refactor(mobile): extract shared iOS harness from the regression tests
The two agent-device suites duplicated ~120 lines of boilerplate (client
bootstrap, simulator prep, snapshot helpers, session cleanup, page-ready poll).
Hoist it into scripts/ios-regressions/harness.ts as createIosHarness(session),
which each test destructures; the test files keep only their own logic (modal's
trigger/scroll helpers, the scrubber's viewport sweep).

Scrubber suite re-run on an iPhone 17 simulator: still passes.
