+++ scripts/ios-regressions/modal.test.ts
@@ -1,119 +1,31 @@
-import type {
-  AgentDeviceClient,
-  AgentDeviceSelectionOptions,
-  CaptureSnapshotResult,
-  SnapshotNode,
-} from "agent-device";
+import type { CaptureSnapshotResult, SnapshotNode } from "agent-device";
 
 import { beforeAll, expect, test } from "bun:test";
-import { execFileSync } from "node:child_process";
-import { existsSync } from "node:fs";
-import { pathToFileURL } from "node:url";
 
-type AgentDeviceModule = typeof import("agent-device");
-
-type IosDeviceOptions = AgentDeviceSelectionOptions & {
-  platform: "ios";
-};
+import { baseUrl, createIosHarness, hookTimeoutMs } from "./harness";
 
 const env = process.env;
-const baseUrl = env.BASE_URL ?? "http://localhost:8000";
 const session = env.SESSION ? `${env.SESSION}-modal` : "zagrajmy-ios-modal-local";
 const targetTitle = env.TARGET_SESSION_TITLE ?? "Przygoda w Mieście Neonów";
 const targetTriggerLabel = env.TARGET_TRIGGER_LABEL ?? `Open details for ${targetTitle}`;
 const eventPath = env.EVENT_PATH ?? "/event/autumn-open/";
 const targetQueryParam = env.TARGET_QUERY_PARAM ?? "session=3";
-const deviceName = env.IOS_DEVICE_NAME ?? "iPhone 16";
-const runtime = env.IOS_RUNTIME;
-const providedUdid = env.UDID;
 const preOpenScrollSteps = Number(env.PRE_OPEN_SCROLL_STEPS ?? "8");
-const hookTimeoutMs = Number(env.IOS_HOOK_TIMEOUT_MS ?? "240000");
 
-const importAgentDevice = async (): Promise<AgentDeviceModule> => {
-  try {
-    return await import("agent-device");
-  } catch (error) {
-    const candidates: string[] = [];
-    try {
-      const npmRoot = execFileSync("npm", ["root", "-g"], { encoding: "utf8" }).trim();
-      candidates.push(`${npmRoot}/agent-device/dist/src/index.js`);
-    } catch {}
-    if (env.HOME) {
-      candidates.push(
-        `${env.HOME}/.bun/install/global/node_modules/agent-device/dist/src/index.js`,
-      );
-    }
-    for (const candidate of candidates) {
-      if (existsSync(candidate)) {
-        return (await import(pathToFileURL(candidate).href)) as AgentDeviceModule;
-      }
-    }
-    throw error;
-  }
-};
-
-const { createAgentDeviceClient } = await importAgentDevice();
-const client: AgentDeviceClient = createAgentDeviceClient({ session });
-
-const deviceOptions: IosDeviceOptions = providedUdid
-  ? { platform: "ios", udid: providedUdid }
-  : { platform: "ios", device: deviceName };
-
-const ensureSimulator = async (): Promise<string> => {
-  if (providedUdid) return providedUdid;
-
-  const result = await client.simulators.ensure({
-    device: deviceName,
-    ...(runtime ? { runtime } : {}),
-    boot: true,
-    reuseExisting: true,
-  });
-  return result.udid;
-};
-
-const takeSnapshot = async (): Promise<CaptureSnapshotResult> =>
-  client.capture.snapshot({ ...deviceOptions, interactiveOnly: true });
-
-const snapshotLabels = async (): Promise<string[]> => {
-  const snapshot = await takeSnapshot();
-  return snapshot.nodes.map((node) => node.label ?? node.value ?? "").filter(Boolean);
-};
+const {
+  client,
+  deviceOptions,
+  takeSnapshot,
+  snapshotLabels,
+  findNodeByLabel,
+  openUrl,
+  prepareDevice,
+  assertPageReady,
+} = await createIosHarness(session);
 
 const hasVisibleText = async (text: string): Promise<boolean> => {
   const labels = await snapshotLabels();
   return labels.some((label) => label.includes(text));
-};
-
-const findNodeByLabel = async (label: string): Promise<SnapshotNode | null> => {
-  const snapshot = await takeSnapshot();
-  return snapshot.nodes.find((node) => node.label === label) ?? null;
-};
-
-const openUrlWithSafari = async (url: string): Promise<void> => {
-  try {
-    await client.apps.open({ ...deviceOptions, app: "Safari", url });
-  } catch (error) {
-    console.warn(
-      "Safari reported a URL open failure; continuing because iOS Simulator can time out after Safari has already loaded the page.",
-      error,
-    );
-  }
-};
-
-const openUrl = async (url: string, udid: string): Promise<void> => {
-  if (!providedUdid) {
-    await openUrlWithSafari(url);
-    return;
-  }
-  try {
-    execFileSync("xcrun", ["simctl", "openurl", udid, url], { stdio: "inherit", timeout: 10000 });
-  } catch (error) {
-    console.warn(
-      "simctl reported a URL open failure; continuing because iOS Simulator can time out before Safari finishes loading.",
-      error,
-    );
-  }
-  await openUrlWithSafari(url);
 };
 
 const clickNodeCenter = async (node: SnapshotNode): Promise<void> => {
@@ -216,59 +128,6 @@ const waitForLabel = async (label: string, timeoutMs: number): Promise<SnapshotN
   return null;
 };
 
-const closeSessionIfPresent = async (name: string): Promise<void> => {
-  try {
-    const sessions = await client.sessions.list();
-    if (!sessions.some((activeSession) => activeSession.name === name)) return;
-
-    console.log(`Taking over existing agent-device session: ${name}`);
-    await client.sessions.close({ session: name });
-  } catch (error) {
-    console.warn(`Could not check or close existing session ${name}:`, error);
-  }
-};
-
-const closeDeviceSessionIfPresent = async (): Promise<void> => {
-  try {
-    const sessions = await client.sessions.list();
-    const activeSession = sessions.find((candidate) => {
-      if (providedUdid) return candidate.device.ios?.udid === providedUdid;
-      return candidate.device.name === deviceName && candidate.device.platform === "ios";
-    });
-    if (!activeSession || activeSession.name === session) return;
-
-    console.log(`Taking over iOS device from existing agent-device session: ${activeSession.name}`);
-    await client.sessions.close({ session: activeSession.name });
-  } catch (error) {
-    console.warn("Could not check or close existing device session:", error);
-  }
-};
-
-const assertEventPageReady = async (url: URL): Promise<void> => {
-  console.log(`Checking local event page at ${url.toString()}...`);
-  const deadline = Date.now() + 60000;
-  let lastError = "no response";
-
-  while (Date.now() < deadline) {
-    try {
-      const response = await fetch(url);
-      const text = await response.text();
-      if (response.ok && text.includes(targetTitle)) return;
-
-      lastError = `HTTP ${response.status}; body starts with ${JSON.stringify(text.slice(0, 160))}`;
-      if (response.status >= 400 && response.status < 500) break;
-    } catch (error) {
-      lastError = error instanceof Error ? error.message : String(error);
-    }
-    await new Promise((resolve) => setTimeout(resolve, 1000));
-  }
-
-  throw new Error(
-    `Local event page is not usable at ${url.toString()} (${lastError}). ` +
-      `Make sure the e2e server is running and serving seeded data for ${targetTitle}.`,
-  );
-};
-
 const eventUrl = new URL(eventPath, baseUrl);
 const modalUrl = new URL(eventUrl);
 for (const [key, value] of new URLSearchParams(targetQueryParam)) {
@@ -282,13 +141,8 @@ let closeIssue: string | null = null;
 let scrollIssue: string | null = null;
 
 beforeAll(async () => {
-  await assertEventPageReady(eventUrl);
-  await closeSessionIfPresent(session);
-  await closeDeviceSessionIfPresent();
-
-  console.log(`Preparing iOS simulator ${providedUdid ?? deviceName}...`);
-  const udid = await ensureSimulator();
-  console.log(`Using simulator UDID: ${udid}`);
+  await assertPageReady(eventUrl, targetTitle);
+  const udid = await prepareDevice();
 
   const initialUrl = openViaScrolledPage ? eventUrl : modalUrl;
   console.log(`Opening Safari at ${initialUrl.toString()}...`);
